0.9 Test

+++ ce1sus/db/classes/attribute.py
@@ -85,7 +85,7 @@ class Attribute(ExtendedLogingInformations, Base):
 
   __bit_code = None
 
-  tlp_level_id = Column('tlp_level_id', Integer(1), default=3, nullable=False)
+  tlp_level_id = Column('tlp_level_id', Integer, default=3, nullable=False)
 
   @property
   def tlp(self):
